part 2 of day 7
This was quite a bit involved and I think the code can be made a lot
easier/nicer. In particular I don't like it that the weight-calculation
is done twice when the code decides it has to recurse. And it does not
work if the root node is the one needing balancing either (i.e. the
list of unbalanced childs is empty). Oh well, for another day.