Solution for day 10
I started out with the idea of generating an endless list of 'theList' to
avoid handling the wrap-around, but couldn't wrap my head around the math
necessary for that. Futile attempt included in commented section.

Went with a simple if/else for handling the wrap-around - a lot easier to
understand for my sunday-morning-brain.