Alternative approach to make it a little easier to follow
This one allows to calculate the position of a scanner for a given delay,
allowing to figure out the list of scanners being at 0 when they're being
passed.

The rest is still brute force trying out wether the list of scanners gets
empty. Unfortunately still the same issue leading to a seemingly endless
calculation.