Micro-Optimization
Drop the len() call needed in findDelay by letting
scannersAtZeroWhenPassingWithDelay also include wether it did an early-exit
because it found the first scanner hit me or not.

The len function turned up in profiling a little (just .260 ms) so this
drops the need for the access ot the length of the list.