Rework the day 13 solution a little
This is easier on the eyes - imo - and still about as fast as my fastest
solution.

Instead of the loop to find the scanners hitting I'm using an iterator
which makes the 'exit on first' vs. 'collect them all' for parts 1 and 2
a lot nicer to write.

I'm not quite satisfied with findDelay, I think that could look nicer with
an iterator generating some endless list too, but I cannot think how to
write that nicely at the moment.