"Fast" solution for part 2 of day 17
Finally, this one still takes quite a bit of time (10s), but its solved
in less than a day. Took me quite a bit until I figured out that the 0
always keeps its leading position and never moves elsewhere so the rest
of the looong list is irrelevant to the solution of part2.