Solution for part1 of day 20 and attempt for part2
The part1 happens to work with my particular input because there's one
particle with a 0-acceleration.

The second part idea was to compare the distances between all particles
of the last run and the current one and check if any of them are 'shrinking'
making a collision probable sometime in the future. If all distances grow
chances are there's no collision anymore.

This does not work if for a particle that 'returns', i.e. where the
acceleration is opposite to its own velocity at the beginning.

It also aborts after 10 iterations for my input at the assert, i.e. the
distances lists are somehow imbalanced.