Fix bug in 'step size' for merging
I ended up taking the wrong grid for anything but the first row because
the startindex wasn't increased by the size of the first row, rather it
always started at 0 in the complete list of grids. I suppose an algorithm
that really 'eats' the input list wouldn't have failed in that way.